+++ .gitignore
@@ -1,6 +1,7 @@
 /testing/__pycache__
 /nnue/__pycache__
-/nnue/_raw
+/nnue/dataset/_raw
 /nnue/parse/parsed_data
+*.dat
 *.o
 *.exe
