+++ nnue/dataset/Makefile
@@ -1,0 +1,24 @@
+CXX = g++
+
+SRC = parse.cpp
+OBJ = parse.o
+EXE = parse.exe
+
+RM = del /q
+
+CPPFLAGS = -I../include
+CXXFLAGS = -Wall -std=gnu++20 -O3
+LDFLAGS = -s -static -static-libstdc++ -static-libgcc
+
+.PHONY: all clean
+
+all: $(EXE)
+
+$(EXE): $(OBJ)
+	$(CXX) $(OBJ) -o $(EXE) $(LDFLAGS)
+
+$(OBJ): $(SRC)
+	$(CXX) $(CPPFLAGS) $(CXXFLAGS) -c $(SRC) -o $(OBJ)
+
+clean:
+	$(RM) $(OBJ) $(EXE)
