+++ nnue/dataset/repo.py
@@ -1,0 +1,3 @@
+REPO_ID = "WillChing01/pingu"
+PATH_IN_REPO = "Pingu_4.0.0"
+REPO_TYPE = "dataset"
